requirements.txt
amqp==5.2.0
asgiref==3.7.2
black==24.2.0
boto3==1.34.57
botocore==1.34.57
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-sendgrid==1.0.1
django-simple-history==3.5.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
kombu==5.2.3
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
paystackapi==2.1.1
pillow==10.2.0
pipenv==2024.0.1
platformdirs==4.2.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
secure==0.3.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.26.3
